updates in admin side for book details
