+++ src/assets/styles/AdminPage.css
@@ -1,13 +1,23 @@
-.adminButtons{
-    position: fixed;
-    margin-top: -30px;
-    width: 100%;
-    background-color: rgb(252, 254, 234);
+.adminMainDiv h2{
+    padding: 20px;
     text-align: center;
-    box-shadow: 1px 1px rgb(201, 193, 106);
+}
+.adminpage{
+    min-height:70vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+.adminButtons{
+    text-align: center;
+    display: flex;
+    flex-direction: column;
 }
 .adminpage button{
-    width: fit-content;
+    width: 200px;
+}
+.adminpage button, .booksList button{
     cursor: pointer;
     padding: 10px;
     background-color: beige;
@@ -16,16 +26,18 @@
     margin: 10px;
     border-radius: 10px;
 }
-.adminpage button:hover{
+.adminpage button:hover, .booksList button:hover{
     background-color: rgb(252, 158, 7); 
     color: white;
 }
 .booksList{
-    padding-top: 40px;
     margin: 30px;
+}
+.booksList h2{
     text-align: center;
 }
 .booksList table, .booksList td, .booksList th{
+    text-align: center;
     border: 1px solid black;
     border-collapse: collapse;
     padding: 10px;
