+++ src/assets/styles/Forms.css
@@ -9,6 +9,9 @@
 }
 .formdiv{
     background-color: beige;
+    border: 1px solid rgb(170, 173, 155);
+    border-radius: 10px;
+    box-shadow: 1px 1px 1px 0px;
     padding: 10px;
     width: 300px;
 }
