+++ src/assets/styles/Header.css
@@ -83,8 +83,9 @@
     background-color: rgb(218, 218, 129);
     width: 120px;
     text-align: center;
-    padding: 10px;
+    padding: 20px;
     border: 1px solid black;
+    border-radius: 10px;
 }
 .profileDropdown button{
     width: 120px;
