+++ src/assets/styles/Profile.css
@@ -25,8 +25,30 @@
     text-decoration: underline;
 }
 .UserProfileContent{
+    margin-top: 30px;
     padding: 10px;
     width: fit-content;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+}
+.UserProfileContent h2{
+    text-align: center;
+}
+.userProfile{
+    margin-top: 20px;
+    padding: 30px;
+    border-radius: 5px;
+    box-shadow: 1px 1px 2px 1px rgb(42, 42, 41);
+}
+.userProfileEditButtons{
+    display: flex;
+    flex-direction: column;
+}
+.userProfileEditButtons button{
+    width: 150px;
+    padding: 5px;
 }
 
 .PasswordDiv{
