+++ src/assets/styles/UserCart.css
@@ -5,7 +5,6 @@
 .emptyCart{
     display: flex;
     flex-direction: column;
-    margin:50px;
     align-items: center;
     justify-content: center;
 }
