+++ src/components/adminPage/viewUserDetails/ViewUserDetails.jsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { Link, useLocation } from "react-router-dom"
 import axios from "axios";
 import { USERDETAILSAPI } from "../../../../public/config/env";
-import '../../../assets/styles/ViewUserDetails.css';
+import '../../../assets/styles/ViewUserdetails.css';
 
 export default function ViewUserDetails() {
     const location = useLocation();
