+++ src/components/userPage/BookRental.jsx
@@ -5,6 +5,7 @@ import BookDetails from '../bookDetails/BookDetails';
 
 export default function BookRental() {
   const auth = useAuth();
+  const data = auth.user;
   const booksDetails = auth.bookDetails;
 
   useEffect(()=>{
@@ -14,6 +15,11 @@ export default function BookRental() {
   return (
 
     <div className="userpage">
+
+      <div className='welcomeDiv'>
+        <h2> Welcome {data.username}!</h2>
+        <img src="https://www.quercusbooks.co.uk/wp-content/uploads/2019/01/Spines-website-asset-new-logo.jpg?w=1920&h=560&crop=1" alt="Book Store image" />
+      </div>
 
       <div className='bookpage'>
           {
