+++ src/components/userPage/changePassword/ChangePassword.jsx
@@ -1,6 +1,5 @@
 import { useState } from "react";
 import { createPortal } from "react-dom";
-import { createRoot } from "react-dom/client";
 
 
 export default function ChangePassword({setEditingNewPassword, password, setChangePassword}) {
